requirements.txt
glob2==0.7
Pillow==8.4.0
PyMuPDF==1.19.2
python-dotenv==0.19.2
toml==0.10.2
